LayoutTests/fast/speech/scripted/speechrecognitionerror-basics.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<script type="text/javascript">
description('Tests the basics of the SpeechRecognitionError interface');

function run() {
    // Check availability of constructor.
    shouldBeTrue("'webkitSpeechRecognitionError' in window");
    shouldBeFalse("webkitSpeechRecognitionError == null");

    finishJSTest();
}

window.onload = run;
window.jsTestIsAsync = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
